elide cache-check for first call to get
get() calls getter.apply(), and getter is initialized to a lambda that replaces getter with the original get() implementation.

This way, the first get() call blindly runs f.apply() while all future calls run it only if needed.

This way, there's no weird edge case when a null output is a correct previous output instead of only being the initial value that indicates that the value hasn't been computed

the state of getter indicates the state of the Manager and directs flow where it needs to go.